# services/api/openapi.json
{
  "components": {
    "schemas": {
      "CashFlowInput": {
        "properties": {
          "amortization_years": {
            "default": 25,
            "maximum": 35.0,
            "minimum": 5.0,
            "title": "Amortization Years",
            "type": "integer"
          },
          "city": {
            "default": "",
            "title": "City",
            "type": "string"
          },
          "closing_costs": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Override; default = land transfer tax estimate + $2,000 legal/inspection",
            "title": "Closing Costs"
          },
          "condo_fee_monthly": {
            "default": 0.0,
            "minimum": 0.0,
            "title": "Condo Fee Monthly",
            "type": "number"
          },
          "down_payment_pct": {
            "default": 20.0,
            "maximum": 100.0,
            "minimum": 5.0,
            "title": "Down Payment Pct",
            "type": "number"
          },
          "insurance_monthly": {
            "default": 0.0,
            "minimum": 0.0,
            "title": "Insurance Monthly",
            "type": "number"
          },
          "interest_rate_pct": {
            "default": 4.5,
            "description": "Fixed rate, e.g. 4.5 for 4.5%",
            "maximum": 25.0,
            "minimum": 0.0,
            "title": "Interest Rate Pct",
            "type": "number"
          },
          "maintenance_pct": {
            "default": 5.0,
            "description": "Share of rent reserved for repairs",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Maintenance Pct",
            "type": "number"
          },
          "management_pct": {
            "default": 0.0,
            "description": "Property manager fee, share of rent",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Management Pct",
            "type": "number"
          },
          "monthly_rent": {
            "minimum": 0.0,
            "title": "Monthly Rent",
            "type": "number"
          },
          "other_income_monthly": {
            "default": 0.0,
            "description": "Parking, storage, laundry\u2026",
            "minimum": 0.0,
            "title": "Other Income Monthly",
            "type": "number"
          },
          "price": {
            "description": "Purchase price, CAD",
            "exclusiveMinimum": 0.0,
            "title": "Price",
            "type": "number"
          },
          "property_tax_annual": {
            "default": 0.0,
            "minimum": 0.0,
            "title": "Property Tax Annual",
            "type": "number"
          },
          "vacancy_pct": {
            "default": 4.0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Vacancy Pct",
            "type": "number"
          }
        },
        "required": [
          "price",
          "monthly_rent"
        ],
        "title": "CashFlowInput",
        "type": "object"
      },
      "CashFlowResult": {
        "properties": {
          "annual_cash_flow": {
            "title": "Annual Cash Flow",
            "type": "number"
          },
          "break_even_rent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Break Even Rent"
          },
          "cap_rate_pct": {
            "title": "Cap Rate Pct",
            "type": "number"
          },
          "cash_invested": {
            "title": "Cash Invested",
            "type": "number"
          },
          "cash_on_cash_pct": {
            "title": "Cash On Cash Pct",
            "type": "number"
          },
          "closing_costs": {
            "title": "Closing Costs",
            "type": "number"
          },
          "closing_costs_note": {
            "title": "Closing Costs Note",
            "type": "string"
          },
          "down_payment": {
            "title": "Down Payment",
            "type": "number"
          },
          "dscr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dscr"
          },
          "effective_monthly_income": {
            "title": "Effective Monthly Income",
            "type": "number"
          },
          "gross_monthly_income": {
            "title": "Gross Monthly Income",
            "type": "number"
          },
          "gross_yield_pct": {
            "title": "Gross Yield Pct",
            "type": "number"
          },
          "insurance_premium": {
            "title": "Insurance Premium",
            "type": "number"
          },
          "monthly_cash_flow": {
            "title": "Monthly Cash Flow",
            "type": "number"
          },
          "monthly_expenses": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Monthly Expenses",
            "type": "object"
          },
          "monthly_mortgage_payment": {
            "title": "Monthly Mortgage Payment",
            "type": "number"
          },
          "monthly_operating_expenses": {
            "title": "Monthly Operating Expenses",
            "type": "number"
          },
          "mortgage_principal": {
            "title": "Mortgage Principal",
            "type": "number"
          },
          "noi_annual": {
            "title": "Noi Annual",
            "type": "number"
          },
          "principal_paydown_year1": {
            "title": "Principal Paydown Year1",
            "type": "number"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "down_payment",
          "insurance_premium",
          "mortgage_principal",
          "monthly_mortgage_payment",
          "closing_costs",
          "closing_costs_note",
          "cash_invested",
          "gross_monthly_income",
          "effective_monthly_income",
          "monthly_expenses",
          "monthly_operating_expenses",
          "noi_annual",
          "monthly_cash_flow",
          "annual_cash_flow",
          "cap_rate_pct",
          "gross_yield_pct",
          "cash_on_cash_pct",
          "dscr",
          "break_even_rent",
          "principal_paydown_year1",
          "warnings"
        ],
        "title": "CashFlowResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HouseInsightsResponse": {
        "properties": {
          "annual_rent": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Annual Rent"
          },
          "cash_flow": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CashFlowSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "gross_yield": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gross Yield"
          },
          "house_id": {
            "title": "House Id",
            "type": "integer"
          },
          "is_synthetic": {
            "default": false,
            "title": "Is Synthetic",
            "type": "boolean"
          },
          "ml": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MlPrediction"
              },
              {
                "type": "null"
              }
            ]
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "net_yield": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Yield"
          },
          "predicted_price": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Predicted Price"
          },
          "price_high": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price High"
          },
          "price_low": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Low"
          },
          "rent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RentEstimate"
              },
              {
                "type": "null"
              }
            ]
          },
          "valuation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Valuation"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "house_id"
        ],
        "title": "HouseInsightsResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Valuation": {
        "properties": {
          "asking_price": {
            "title": "Asking Price",
            "type": "integer"
          },
          "basis": {
            "default": "active asking prices",
            "title": "Basis",
            "type": "string"
          },
          "comps": {
            "items": {
              "$ref": "#/components/schemas/Comp"
            },
            "title": "Comps",
            "type": "array"
          },
          "confidence": {
            "title": "Confidence",
            "type": "string"
          },
          "delta_pct": {
            "title": "Delta Pct",
            "type": "number"
          },
          "fair_value": {
            "title": "Fair Value",
            "type": "integer"
          },
          "fair_value_high": {
            "title": "Fair Value High",
            "type": "integer"
          },
          "fair_value_low": {
            "title": "Fair Value Low",
            "type": "integer"
          },
          "median_price_per_sqft": {
            "title": "Median Price Per Sqft",
            "type": "number"
          },
          "method": {
            "default": "comps-v1",
            "title": "Method",
            "type": "string"
          },
          "radius_m": {
            "title": "Radius M",
            "type": "integer"
          },
          "subject_price_per_sqft": {
            "title": "Subject Price Per Sqft",
            "type": "number"
          },
          "verdict": {
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "asking_price",
          "fair_value",
          "fair_value_low",
          "fair_value_high",
          "median_price_per_sqft",
          "subject_price_per_sqft",
          "delta_pct",
          "verdict",
          "confidence",
          "radius_m",
          "comps"
        ],
        "title": "Valuation",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Canadian residential investment analytics: listings, comparables, cash flow, markets.",
    "title": "NeighborIQ API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cashflow": {
      "post": {
        "description": "Stateless cash-flow projection for user-edited assumptions.",
        "operationId": "post_cash_flow_api_v1_cashflow_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashFlowInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashFlowResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Cash Flow",
        "tags": [
          "insights"
        ]
      }
    },
...
    "/api/v1/houses/{house_id}/insights": {
      "get": {
        "description": "Everything an investor needs for one listing: comps-based fair value, a rent\nestimate, and a cash-flow projection at default assumptions (recompute with\nPOST /api/v1/cashflow after editing them).",
        "operationId": "get_house_insights_api_v1_houses__house_id__insights_get",
        "parameters": [
          {
            "in": "path",
            "name": "house_id",
            "required": true,
            "schema": {
              "title": "House Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseInsightsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get House Insights",
        "tags": [
          "insights"
        ]
      }
    },
...
    "/api/v1/houses/{house_id}/valuation": {
      "get": {
        "description": "Comparable-listing valuation only. null = not enough comparable listings.",
        "operationId": "get_house_valuation_api_v1_houses__house_id__valuation_get",
        "parameters": [
          {
            "in": "path",
            "name": "house_id",
            "required": true,
            "schema": {
              "title": "House Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Valuation"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get House Valuation Api V1 Houses  House Id  Valuation Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get House Valuation",
        "tags": [
          "insights"
        ]
      }
    },
    "/api/v1/markets": {
      "get": {
        "description": "Per-city market snapshot computed from active listings.",
        "operationId": "list_markets_api_v1_markets_get",
        "parameters": [
          {
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MarketSummary"
                  },
                  "title": "Response List Markets Api V1 Markets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Markets",
        "tags": [
          "insights"
        ]
      }
    },
    "/api/v1/markets/{city}/points": {
      "get": {
        "description": "Compact per-listing values for map aggregation (hexagons, heat, extrusions):\ncolumns = [id, lat, lon, price, price_per_sqft, gross_yield_pct, cap_rate_pct,\nprice_cut_pct, days_on_market]. Arrays, not objects, to keep the payload small.",
        "operationId": "market_points_api_v1_markets__city__points_get",
        "parameters": [
          {
            "in": "path",
            "name": "city",
            "required": true,
            "schema": {
              "title": "City",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5000,
              "maximum": 20000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Market Points",
        "tags": [
          "insights"
        ]
      }
    },
    "/api/v1/neighborhoods/{city}/{region}/analysis": {
      "get": {
        "operationId": "get_neighborhood_analysis_api_v1_neighborhoods__city___region__analysis_get",
        "parameters": [
          {
            "in": "path",
            "name": "city",
            "required": true,
            "schema": {
              "title": "City",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "region",
            "required": true,
            "schema": {
              "title": "Region",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeighborhoodAnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Neighborhood Analysis",
        "tags": [
          "insights"
        ]
      }
    },
...
  }
}